move the class of apply_augmention into drivaernet_dataset.py and modify the markdown file about apply_augmention
